+++ cortex-m-rt/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v0.3.0] - 2017-07-07
+
 ### Added
 
 - A `default_handler!` macro to override the default exception handler.
@@ -118,7 +120,8 @@ section                size        addr
 
 Initial release
 
-[Unreleased]: https://github.com/japaric/cortex-m-rt/compare/v0.2.4...HEAD
+[Unreleased]: https://github.com/japaric/cortex-m-rt/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/japaric/cortex-m-rt/compare/v0.2.4...v0.3.0
 [v0.2.4]: https://github.com/japaric/cortex-m-rt/compare/v0.2.3...v0.2.4
 [v0.2.3]: https://github.com/japaric/cortex-m-rt/compare/v0.2.2...v0.2.3
 [v0.2.2]: https://github.com/japaric/cortex-m-rt/compare/v0.2.1...v0.2.2